addressing pr comment issues
